Add duration to VzaarApi::Video
When looking up a video, Vzaar's API response JSON includes the duration
of a video, among other metadata.

Currently this is not exposed in VzaarApi::Video. It is possible to
access it with `JSON.parse(video.to_json)['duration']` as a work around.

This commit adds a read-only `.duration` attribute to VzaarApi::Video.